+++ CHANGELOG.rst
@@ -6,6 +6,14 @@ fourfront
 Change Log
 ----------
 
+5.1.4
+=====
+
+`PR Vitessce upgrade  <https://github.com/4dn-dcic/fourfront/pull/1792>`_
+
+* upgrade from 1.1.20 to 2.0.3
+
+
 5.1.3
 =====